apps/api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.18
docling==2.14.0
python-dotenv==1.0.1
boto3==1.35.0
httpx==0.28.1
pydantic>=2.11.0,<3.0.0
pydantic-settings>=2.7.0,<3.0.0
weasyprint==63.1
markdown==3.7
python-docx>=1.1.0
jinja2==3.1.4
python-jose[cryptography]==3.3.0
bcrypt>=4.0.0
email-validator>=2.0.0
openpyxl>=3.1.0
ebooklib>=0.18
pyyaml>=6.0
